// myapp/src/services/ant-design-pro/typings.d.ts
// @ts-ignore
/* eslint-disable */

import {RcFile} from "antd/es/upload";

declare namespace API {
  type CurrentUser = {
    id: number;
    username?: string;
    userAccount: string;
    avatarUrl?: string;
    gender?: number;
    phone?: string;
    email?: string;
    userStatus?: number;
    userRole?: number;
    createTime?: Date;
    unreadCount?: number;
  };

  type LoginResult = {
    status?: string;
    type?: string;
    currentAuthority?: string;
  };

  type PageParams = {
    current?: number;
    pageSize?: number;
  };

  type RuleListItem = {
    key?: number;
    disabled?: boolean;
    href?: string;
    avatar?: string;
    name?: string;
    owner?: string;
    desc?: string;
    callNo?: number;
    status?: number;
    updatedAt?: string;
    createdAt?: string;
    progress?: number;
  };

  type RuleList = {
    data?: RuleListItem[];
    /** 列表的内容总数 */
    total?: number;
    success?: boolean;
  };

  type FakeCaptcha = {
    code?: number;
    status?: string;
  };

  type LoginParams = {
    userAccount?: string;
    userPassword?: string;
    autoLogin?: boolean;
    type?: string;
  };

  type RegisterParams = {
    userName?: string;
    userPassword?: string;
    checkPassword?: string;
    gender?: number;
    type?: string;
  };

  type SearchParams = {
    userName?: string;
    userAccount?: string;
    gender?: number;
    phone?: string;
    email?: string;
  }
  type UpdateInfoParams = {
    username?: string;
    userAccount?: string;
    gender?: number;
    phone?: string;
    email?: string;
    avatar?: string;
  }
  /**
   * 通用返回类
   */
  type Response<T> = {
    code: string;
    data: T;
    message: string;
    description?: string;
  }

  type ErrorResponse = {
    /** 业务约定的错误码 */
    errorCode: string;
    /** 业务上的错误信息 */
    errorMessage?: string;
    /** 业务上的请求是否成功 */
    success?: boolean;
  };

  type NoticeIconList = {
    data?: NoticeIconItem[];
    /** 列表的内容总数 */
    total?: number;
    success?: boolean;
  };

  type NoticeIconItemType = 'notification' | 'message' | 'event';

  type NoticeIconItem = {
    id?: string;
    extra?: string;
    key?: string;
    read?: boolean;
    avatar?: string;
    title?: string;
    status?: string;
    datetime?: string;
    description?: string;
    type?: NoticeIconItemType;
  };
}